+++ packages/framework/tests/Feature/Services/CheckSumServiceTest.php
@@ -3,17 +3,17 @@
 namespace Hyde\Framework\Testing\Feature\Services;
 
 use Hyde\Framework\Hyde;
-use Hyde\Framework\Services\ViewDiffService;
+use Hyde\Framework\Services\CheckSumService;
 use Hyde\Testing\TestCase;
 
 /**
- * @covers \Hyde\Framework\Services\ViewDiffService
+ * @covers \Hyde\Framework\Services\CheckSumService
  */
-class ViewDiffServiceTest extends TestCase
+class CheckSumServiceTest extends TestCase
 {
     public function test_get_filecache()
     {
-        $fileCacheService = new ViewDiffService();
+        $fileCacheService = new CheckSumService();
         $fileCache = $fileCacheService->getFilecache();
 
         $this->assertIsArray($fileCache);
@@ -24,7 +24,7 @@ class ViewDiffServiceTest extends TestCase
 
     public function test_get_checksums()
     {
-        $fileCacheService = new ViewDiffService();
+        $fileCacheService = new CheckSumService();
         $checksums = $fileCacheService->getChecksums();
 
         $this->assertIsArray($checksums);
@@ -33,69 +33,69 @@ class ViewDiffServiceTest extends TestCase
 
     public function test_checksum_matches_any()
     {
-        $fileCacheService = new ViewDiffService();
+        $fileCacheService = new CheckSumService();
 
-        $this->assertTrue($fileCacheService->checksumMatchesAny(ViewDiffService::unixsumFile(
+        $this->assertTrue($fileCacheService->checksumMatchesAny(CheckSumService::unixsumFile(
             Hyde::vendorPath('resources/views/layouts/app.blade.php'))
         ));
     }
 
     public function test_checksum_matches_any_false()
     {
-        $fileCacheService = new ViewDiffService();
+        $fileCacheService = new CheckSumService();
 
-        $this->assertFalse($fileCacheService->checksumMatchesAny(ViewDiffService::unixsum(
+        $this->assertFalse($fileCacheService->checksumMatchesAny(CheckSumService::unixsum(
             'foo'
         )));
     }
 
     public function test_method_returns_string()
     {
-        $this->assertIsString(ViewDiffService::unixsum('foo'));
+        $this->assertIsString(CheckSumService::unixsum('foo'));
     }
 
     public function test_method_returns_string_with_length_of_32()
     {
-        $this->assertEquals(32, strlen(ViewDiffService::unixsum('foo')));
+        $this->assertEquals(32, strlen(CheckSumService::unixsum('foo')));
     }
 
     public function test_method_returns_string_matching_expected_format()
     {
-        $this->assertMatchesRegularExpression('/^[a-f0-9]{32}$/', ViewDiffService::unixsum('foo'));
+        $this->assertMatchesRegularExpression('/^[a-f0-9]{32}$/', CheckSumService::unixsum('foo'));
     }
 
     public function test_method_returns_same_value_for_same_string_using_normal_method()
     {
-        $this->assertEquals(md5('foo'), ViewDiffService::unixsum('foo'));
+        $this->assertEquals(md5('foo'), CheckSumService::unixsum('foo'));
     }
 
     public function test_method_returns_different_value_for_different_string()
     {
-        $this->assertNotEquals(ViewDiffService::unixsum('foo'), ViewDiffService::unixsum('bar'));
+        $this->assertNotEquals(CheckSumService::unixsum('foo'), CheckSumService::unixsum('bar'));
     }
 
     public function test_function_is_case_sensitive()
     {
-        $this->assertNotEquals(ViewDiffService::unixsum('foo'), ViewDiffService::unixsum('FOO'));
+        $this->assertNotEquals(CheckSumService::unixsum('foo'), CheckSumService::unixsum('FOO'));
     }
 
     public function test_function_is_space_sensitive()
     {
-        $this->assertNotEquals(ViewDiffService::unixsum(' foo '), ViewDiffService::unixsum('foo'));
+        $this->assertNotEquals(CheckSumService::unixsum(' foo '), CheckSumService::unixsum('foo'));
     }
 
     public function test_method_returns_same_value_regardless_of_end_of_line_sequence()
     {
-        $this->assertEquals(ViewDiffService::unixsum('foo'), ViewDiffService::unixsum('foo'));
-        $this->assertEquals(ViewDiffService::unixsum("foo\n"), ViewDiffService::unixsum("foo\n"));
-        $this->assertEquals(ViewDiffService::unixsum("foo\n"), ViewDiffService::unixsum("foo\r"));
-        $this->assertEquals(ViewDiffService::unixsum("foo\n"), ViewDiffService::unixsum("foo\r\n"));
+        $this->assertEquals(CheckSumService::unixsum('foo'), CheckSumService::unixsum('foo'));
+        $this->assertEquals(CheckSumService::unixsum("foo\n"), CheckSumService::unixsum("foo\n"));
+        $this->assertEquals(CheckSumService::unixsum("foo\n"), CheckSumService::unixsum("foo\r"));
+        $this->assertEquals(CheckSumService::unixsum("foo\n"), CheckSumService::unixsum("foo\r\n"));
     }
 
     public function test_method_returns_same_value_for_string_with_mixed_end_of_line_sequences()
     {
-        $this->assertEquals(ViewDiffService::unixsum("foo\nbar\r\nbaz\r\n"),
-            ViewDiffService::unixsum("foo\nbar\nbaz\n"));
+        $this->assertEquals(CheckSumService::unixsum("foo\nbar\r\nbaz\r\n"),
+            CheckSumService::unixsum("foo\nbar\nbaz\n"));
     }
 
     public function test_method_returns_same_value_when_loaded_from_file()
@@ -104,7 +104,7 @@ class ViewDiffServiceTest extends TestCase
         $file = tempnam(sys_get_temp_dir(), 'foo');
         file_put_contents($file, $string);
 
-        $this->assertEquals(ViewDiffService::unixsum($string), ViewDiffService::unixsum(file_get_contents($file)));
+        $this->assertEquals(CheckSumService::unixsum($string), CheckSumService::unixsum(file_get_contents($file)));
 
         unlink($file);
     }
@@ -115,7 +115,7 @@ class ViewDiffServiceTest extends TestCase
         $file = tempnam(sys_get_temp_dir(), 'foo');
         file_put_contents($file, $string);
 
-        $this->assertEquals(ViewDiffService::unixsum($string), ViewDiffService::unixsumFile($file));
+        $this->assertEquals(CheckSumService::unixsum($string), CheckSumService::unixsumFile($file));
 
         unlink($file);
     }